+++ networkx/algorithms/tests/test_covering.py
@@ -83,25 +83,3 @@ class TestIsEdgeCover:
         assert nx.is_edge_cover(G, {(0, 1), (1, 0)})
         assert nx.is_edge_cover(G, {(0, 1)})
         assert not nx.is_edge_cover(G, {(0, 0)})
-
-
-class TestIsVertexCover:
-    """Tests for :func:`networkx.algorithms.is_vertex_cover`"""
-
-    def test_empty_graph(self):
-        G = nx.Graph()
-        assert nx.is_vertex_cover(G, set())
-
-    def test_graph_with_loop(self):
-        G = nx.Graph()
-        G.add_edge(1, 1)
-        assert nx.is_vertex_cover(G, {1})
-
-    def test_graph_double_edge(self):
-        G = nx.Graph()
-        G.add_edge(0, 1)
-        G.add_edge(1, 2)
-
-        assert nx.is_vertex_cover(G, {1})
-        assert nx.is_vertex_cover(G, {0, 2})
-        assert not nx.is_vertex_cover(G, {0})
